+++ packages/browser-wallet/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+-   Display contract address of issuer in verifiable credential details.
+
 ### Fixed
 
+-   Incorrect verifiable presentations created, due to incorrect identity/identityProviderIndex used.
+-   Wallet crashing when showing a proof request, while having a verifiable credential that is not yet on chain (or we otherwise fail to retrieve the status)
 -   Show verifiable credentials in overview before they are put on chain.
 
 ## 1.1.1